requirements.txt
# Core ML Dependencies
torch>=2.0.0,<3.0.0
torchvision>=0.15.0,<1.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.0.0,<2.0.0

# Image Processing
opencv-python>=4.5.0,<5.0.0
Pillow>=8.3.0,<11.0.0

# Visualization and Progress
matplotlib>=3.5.0,<4.0.0
tqdm>=4.62.0,<5.0.0

# Production Dependencies
pathlib2>=2.3.0,<3.0.0
psutil>=5.8.0,<6.0.0
pyyaml>=6.0,<7.0
click>=8.0.0,<9.0.0
colorama>=0.4.4,<1.0.0
rich>=12.0.0,<13.0.0

# Logging and Monitoring
structlog>=22.0.0,<23.0.0
python-json-logger>=2.0.0,<3.0.0

# Testing and Development (optional)
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
black>=22.0.0,<23.0.0
flake8>=5.0.0,<6.0.0 